move bufferOps to core & merge with jsonOps
