+++ core/lib.rs
@@ -15,6 +15,8 @@ mod module_specifier;
 mod modules;
 mod normalize_path;
 mod ops;
+mod ops_buffer;
+mod ops_json;
 pub mod plugin_api;
 mod resources;
 mod runtime;
@@ -58,8 +60,6 @@ pub use crate::modules::ModuleSourceFuture;
 pub use crate::modules::NoopModuleLoader;
 pub use crate::modules::RecursiveModuleLoad;
 pub use crate::normalize_path::normalize_path;
-pub use crate::ops::json_op_async;
-pub use crate::ops::json_op_sync;
 pub use crate::ops::op_close;
 pub use crate::ops::op_resources;
 pub use crate::ops::Op;
@@ -68,6 +68,11 @@ pub use crate::ops::OpFn;
 pub use crate::ops::OpId;
 pub use crate::ops::OpState;
 pub use crate::ops::OpTable;
+pub use crate::ops_buffer::buffer_op_async;
+pub use crate::ops_buffer::buffer_op_sync;
+pub use crate::ops_buffer::ValueOrVector;
+pub use crate::ops_json::json_op_async;
+pub use crate::ops_json::json_op_sync;
 pub use crate::resources::Resource;
 pub use crate::resources::ResourceId;
 pub use crate::resources::ResourceTable;
